# [perf] Set all:false in stats.toJson call